add todo query
