+++ pkg/graph/mutation.resolvers.go
@@ -4,14 +4,24 @@ package graph
 // will be copied through when generating and any unknown code will be moved to the end.
 
 import (
+	"context"
+	"log"
+
 	"github.com/y-mabuchi/torasemi-todo-api/pkg/graph/generated"
+	"github.com/y-mabuchi/torasemi-todo-api/pkg/graph/model"
 )
+
+func (r *mutationResolver) CreateUser(ctx context.Context, input *model.CreateUserInput) (*model.User, error) {
+	user, err := r.repo.CreateUser(ctx, input)
+	if err != nil {
+		log.Printf("action=r.repo.CreateUser, status=error: %v", err)
+		return nil, err
+	}
+
+	return &model.User{User: user}, nil
+}
 
 // Mutation returns generated.MutationResolver implementation.
 func (r *Resolver) Mutation() generated.MutationResolver { return &mutationResolver{r} }
 
-// Query returns generated.QueryResolver implementation.
-func (r *Resolver) Query() generated.QueryResolver { return &queryResolver{r} }
-
 type mutationResolver struct{ *Resolver }
-type queryResolver struct{ *Resolver }
