+++ schema/base.graphql
@@ -1,3 +1,1 @@
 scalar Time
-type Query
-type Mutation
