+++ schema/query.graphql
@@ -1,0 +1,9 @@
+type Query {
+    # users
+    allUsers: [User]!
+    user(id: ID!): User!
+
+    # todos
+    allTodos: [Todo]!
+    todo(id: ID!): Todo!
+}
